Some stuff...
I probably should have commited before, but everything was mixed up together. Added mongoose to connect to a database. That database not only holds users, but also the session data. That way when the server is restarted the session data is still saved. Also got the password hashing applied to the user schema to improve productivity. Lastly, I added a README to keep track of my progress