That took waaaaaay more effort than I thought it would: Update a record
Most of the problems arose when I tried to update an array of records. I'm sure I missed somethign that would have made it quick, but I learned a lot of things along the way about php and how to debug. Hope it's not that awful for the rest of this.