Resizing setup
Resizing is proving to be rather difficult. I decided to first capture getting the cursor to change on the proper edge and therefore set it up to resize in that direction. Also set up the function to do the actual resizing, but it's just a start. I'll probably employ the same technique I did with the drag (move the mouse to it if we're 'resizing')