Got the basics of resizing to work
It's only the bottom, but it's something. My problem is the same one that I had with dragging, but this time since it's resizing we might have to go up a level (which is what I'm trying to avoid). The cursor once it leaves the box it leaves the 'area of effect' and the resizing stops. Will see if I can find a work around, but till then this might be just what it is