Got the resize working
There were some tricks done to get it to work. One in particular was that when you're resizing you need to +1 do the width to keep the cursor on the block (works out rather nice). That doesn't work with the lefthand side since you need to both resize AND move the block over. For corners it was simple the combination of the two