Implementing the z-indexing
I did this for the last commit (I think...) but it needed some fine tuning. Mostly setting the zIndex of the items to being the highest (in order) when clicked. That does mean that if they click between them 9998 times there will be a problem, but I think that's a bug I would live with.